clean up

+++ src/app/controllers/stock-list.component.ts
@@ -13,7 +13,7 @@ import { Favorite } from '../models/favorite';
 
 export class StockListComponent implements OnInit, OnDestroy{
 
-  stockList:Stock[] = [];
+  stocks = {active:true, list:[]};
   private favoriteStream;
   constructor(private stockService:StockService, private auth: AuthGuard,  private favoriteService: FavoriteService){}
 
@@ -23,9 +23,9 @@ export class StockListComponent implements OnInit, OnDestroy{
       .then((values)=>{
         values[0].json().forEach((stock) => {
           if(values[1].json().length && values[1].json().find((favorite) => favorite.stockId === stock.id)) {
-            this.stockList.push(new Stock(stock.name, stock.symbol, stock.exchange, stock.id, true));
+            this.stocks.list.push(new Stock(stock.name, stock.symbol, stock.exchange, stock.id, true));
           } else {
-            this.stockList.push(new Stock(stock.name, stock.symbol, stock.exchange, stock.id, false));
+            this.stocks.list.push(new Stock(stock.name, stock.symbol, stock.exchange, stock.id, false));
           }
         });
       })
@@ -35,6 +35,7 @@ export class StockListComponent implements OnInit, OnDestroy{
       // Subscribe for updates of favorites
       this.favoriteStream = this.favoriteService.favoriteStream()
         .subscribe((favorite: Favorite )=> {
+          console.log("update");
             this.updateFavoriteList(favorite);
         });
   }
@@ -44,7 +45,7 @@ export class StockListComponent implements OnInit, OnDestroy{
   }
 
   updateFavoriteList(favorite: Favorite) {
-    this.stockList.map((stock) => {
+    this.stocks.list.map((stock) => {
       if(favorite.stockId === stock.id) {
         stock.favorite = favorite.isFavorite;
         return stock;
@@ -53,23 +54,29 @@ export class StockListComponent implements OnInit, OnDestroy{
   }
 
   alterFavorite(stock: Stock) {
-    if(stock.favorite) {
-      this.favoriteService.removeFavorite(stock.id)
-        .then((res) => {
-          console.log("favorite removed");
-        })
-        .catch((err) => {
-          console.log(err);
-        });
+    if(this.stocks.active){
+      this.stocks.active = false;
+      if(stock.favorite) {
+        this.favoriteService.removeFavorite(stock.id)
+          .then((res) => {
+            this.stocks.active = true;
+            console.log("favorite removed");
+          })
+          .catch((err) => {
+            console.log(err);
+          });
 
-    } else {
-      this.favoriteService.addFavorite(stock.id)
-        .then((data) => {
-          console.log("favorite added");
-        })
-        .catch((err) => {
-          console.log(err);
-        });
+      } else {
+        this.favoriteService.addFavorite(stock.id)
+          .then((data) => {
+            this.stocks.active = true;
+            console.log("favorite added");
+          })
+          .catch((err) => {
+            console.log(err);
+          });
+      }
     }
+
   }
 }
